Fix Visual Studio compiler warnings (64 bit) (#2189)
Some warnings are specific to 64 bit systems. Compiled with Visual
Studio 2022 on Windows 11 x64.

+++ libarchive/test/test_read_format_rar5.c
@@ -843,7 +843,7 @@ DEFINE_TEST(test_read_format_rar5_block_by_block)
 	struct archive_entry *ae;
 	struct archive *a;
 	uint8_t buf[173];
-	int bytes_read;
+	ssize_t bytes_read;
 	uint32_t computed_crc = 0;
 
 	extract_reference_file("test_read_format_rar5_compressed.rar");
